requirements.txt
blinker==1.6.3
certifi==2020.12.5
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.10.0
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.43.1
imageio==2.31.5
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.3.2
networkx==3.1
numpy==1.19.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.4.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
six==1.16.0
threadpoolctl==3.2.0
tifffile==2023.7.10
Werkzeug==3.0.0
zipp==3.17.0
